+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Fix to installation script: missing `sudo` to remove old library, not observing the GOPJRT_NOSUDO request.
 * Fixed github test action `go.yaml`.
+* Added dtypes.IsUnsigned.
 
 # v0.4.5